Move Updates 1
Begun adding in all moves that effect the damage calculation in someway.

+++ demo/Battle Calculator/src/MoveClass.java
@@ -31,7 +31,7 @@ public class MoveClass implements Serializable {
         return category;
     }
 
-    public void setBase(int base) {
+    public void setBase(double base) {
         this.base = base;
     }
